services/mcp-research/requirements.txt
# Sophia AI Research MCP - Requirements
# Core FastAPI and async support
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2

# Database drivers
asyncpg==0.29.0

# HTTP client for external APIs
httpx==0.27.0

# Vector database - using Weaviate instead of Qdrant
weaviate-client==4.4.1

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# Utilities
python-dotenv==1.0.1
python-json-logger==2.0.7

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.1